+++ packages/types/src/definitions/runtime/Extratimer.type.ts
@@ -1,0 +1,7 @@
+export type SimplePlayback = 'play' | 'pause' | 'stop';
+
+export type SimpleTimerState = {
+    duration: number;
+    current: number;
+    playback: SimplePlayback;
+}
